+++ http-client-tls/test/Spec.hs
@@ -9,7 +9,5 @@ main :: IO ()
 main = hspec $ do
     it "make a TLS connection" $ do
         manager <- newManager tlsManagerSettings
-        withResponse "https://httpbin.org/status/418"
-            { checkStatus = \_ _ _ -> Nothing
-            } manager $ \res -> do
+        withResponse "https://httpbin.org/status/418" manager $ \res ->
             responseStatus res `shouldBe` status418