+++ http-client/ChangeLog.md
@@ -3,6 +3,7 @@
 * Remove `instance Default Request`
 * Modify `instance IsString Request` to use `parseRequest` instead of `parseUrlThrow`
 * Clean up the `HttpException` constructors
+* Rename `checkStatus` to `checkResponse` and modify type
 
 ## 0.4.30